proto0001/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
psycopg2-binary==2.9.9
bcrypt==4.1.2
PyJWT==2.8.0
python-dotenv==1.0.0
openai==1.6.1